requirements.txt
streamlit
chromadb
python-dotenv
mistralai
pymupdf
pandas